Adjust phc2sys startup to wait until after ntpdate runs.
For applications using PTP synchronization, the setup instructions show how
to run phc2sys at startup, which is necessary to keep the hardware clock in
the NIC synchronized with the system clock.  The setup instructions rely on
the default ntpdate configuration to initialize the system clock with the
current real-world time. Before, the suggested phc2sys startup script doesn't
properly wait for ntpdate to complete, so that when ntpdate does finish,
phc2sys sees a large jump in the system time.  Because phc2sys slowly adjusts
the NIC clock, the time published by ptp4l will then not be synchronized with
the system clock, and could take a very long time to do so.  This has the
effect of publishing a time with ingress data that is out-of-sync with the
system time.  This change request fixes the documentation to include a
startup instruction that waits for ntpdate to finish before running phc2sys.

See https://github.com/nvidia-holoscan/holoscan-sensor-bridge/issues/4.